Remove remaining code from typecheck
