--- lang/elaborator/src/typechecker/exprs/typecheck.rs
@@ -1,145 +1,0 @@
-//! Bidirectional type checker
-
-use std::rc::Rc;
-
-use codespan::Span;
-
-use crate::normalizer::env::ToEnv;
-use crate::normalizer::normalize::Normalize;
-use miette_util::ToMiette;
-use syntax::ast::*;
-use syntax::common::*;
-use syntax::ctx::values::Binder;
-use syntax::ctx::{BindContext, BindElem};
-
-use super::super::ctx::*;
-use super::CheckInfer;
-use crate::result::TypeError;
-
-// Other
-//
-//
-
-pub trait CheckTelescope {
-    type Target;
-
-    fn check_telescope<T, F: FnOnce(&mut Ctx, Self::Target) -> Result<T, TypeError>>(
-        &self,
-        prg: &Module,
-        name: &str,
-        ctx: &mut Ctx,
-        params: &Telescope,
-        f: F,
-        span: Option<Span>,
-    ) -> Result<T, TypeError>;
-}
-
-pub trait InferTelescope {
-    type Target;
-
-    fn infer_telescope<T, F: FnOnce(&mut Ctx, Self::Target) -> Result<T, TypeError>>(
-        &self,
-        prg: &Module,
-        ctx: &mut Ctx,
-        f: F,
-    ) -> Result<T, TypeError>;
-}
-
-impl CheckTelescope for TelescopeInst {
-    type Target = TelescopeInst;
-
-    fn check_telescope<T, F: FnOnce(&mut Ctx, Self::Target) -> Result<T, TypeError>>(
-        &self,
-        prg: &Module,
-        name: &str,
-        ctx: &mut Ctx,
-        param_types: &Telescope,
-        f: F,
-        span: Option<Span>,
-    ) -> Result<T, TypeError> {
-        let Telescope { params: param_types } = param_types;
-        let TelescopeInst { params } = self;
-
-        if params.len() != param_types.len() {
-            return Err(TypeError::ArgLenMismatch {
-                name: name.to_owned(),
-                expected: param_types.len(),
-                actual: params.len(),
-                span: span.to_miette(),
-            });
-        }
-
-        let iter = params.iter().zip(param_types);
-
-        ctx.bind_fold_failable(
-            iter,
-            vec![],
-            |ctx, params_out, (param_actual, param_expected)| {
-                let ParamInst { span, name, .. } = param_actual;
-                let Param { typ, .. } = param_expected;
-                let typ_out = typ.check(prg, ctx, Rc::new(TypeUniv::new().into()))?;
-                let typ_nf = typ.normalize(prg, &mut ctx.env())?;
-                let mut params_out = params_out;
-                let param_out = ParamInst {
-                    span: *span,
-                    info: Some(typ_nf.clone()),
-                    name: name.clone(),
-                    typ: typ_out.into(),
-                };
-                params_out.push(param_out);
-                let elem = Binder { name: param_actual.name.clone(), typ: typ_nf };
-                Result::<_, TypeError>::Ok(BindElem { elem, ret: params_out })
-            },
-            |ctx, params| f(ctx, TelescopeInst { params }),
-        )?
-    }
-}
-
-impl InferTelescope for Telescope {
-    type Target = Telescope;
-
-    fn infer_telescope<T, F: FnOnce(&mut Ctx, Self::Target) -> Result<T, TypeError>>(
-        &self,
-        prg: &Module,
-        ctx: &mut Ctx,
-        f: F,
-    ) -> Result<T, TypeError> {
-        let Telescope { params } = self;
-
-        ctx.bind_fold_failable(
-            params.iter(),
-            vec![],
-            |ctx, mut params_out, param| {
-                let Param { typ, name } = param;
-                let typ_out = typ.check(prg, ctx, Rc::new(TypeUniv::new().into()))?;
-                let typ_nf = typ.normalize(prg, &mut ctx.env())?;
-                let param_out = Param { name: name.clone(), typ: typ_out };
-                params_out.push(param_out);
-                let elem = Binder { name: param.name.clone(), typ: typ_nf };
-                Result::<_, TypeError>::Ok(BindElem { elem, ret: params_out })
-            },
-            |ctx, params| f(ctx, Telescope { params }),
-        )?
-    }
-}
-
-impl InferTelescope for SelfParam {
-    type Target = SelfParam;
-
-    fn infer_telescope<T, F: FnOnce(&mut Ctx, Self::Target) -> Result<T, TypeError>>(
-        &self,
-        prg: &Module,
-        ctx: &mut Ctx,
-        f: F,
-    ) -> Result<T, TypeError> {
-        let SelfParam { info, name, typ } = self;
-
-        let typ_nf = typ.to_exp().normalize(prg, &mut ctx.env())?;
-        let typ_out = typ.infer(prg, ctx)?;
-        let param_out = SelfParam { info: *info, name: name.clone(), typ: typ_out };
-        let elem = Binder { name: name.clone().unwrap_or_default(), typ: typ_nf };
-
-        // We need to shift the self parameter type here because we treat it as a 1-element telescope
-        ctx.bind_single(&elem.shift((1, 0)), |ctx| f(ctx, param_out))
-    }
-}
